Move `walk_files` to fs.py
